# Change isKnownService() cache to reduce memory allocation