--- a/render/theinternet.go
+++ b/render/theinternet.go
@@ -3,8 +3,9 @@ package render
 import (
 	"regexp"
 	"strings"
+	"time"
 
-	"github.com/bluele/gcache"
+	gocache "github.com/patrickmn/go-cache"
 
 	"github.com/weaveworks/scope/probe/host"
 	"github.com/weaveworks/scope/report"
@@ -36,28 +37,24 @@ var (
 
 	// Memoization for isKnownService.
 	//
-	// The 10000 comes from the observation that large reports contain
-	// hundreds of names, and in a multi-tenant context we want to be
-	// able to render a few dozen reports concurrently. Also, unlike
-	// memoization in the reducers, which is keyed on reports, this
 	// cache is effective when rendering multiple reports from the
 	// same cluster of probes, e.g. from different points in time,
 	// since names tend to change infrequently.
 	//
 	// Since names are generally <50 bytes, this shouldn't weight in
 	// at more than a few MB of memory.
-	knownServiceCache = gcache.New(10000).ARC().Build()
+	knownServiceCache = gocache.New(10*time.Minute, 10*time.Minute)
 )
 
 // TODO: Make it user-customizable https://github.com/weaveworks/scope/issues/1876
 // NB: this is a hotspot in rendering performance.
 func isKnownService(hostname string) bool {
-	if v, err := knownServiceCache.Get(hostname); err == nil {
+	if v, found := knownServiceCache.Get(hostname); found {
 		return v.(bool)
 	}
 
 	known := knownServiceMatcher.MatchString(hostname) && !knownServiceExcluder.MatchString(hostname)
-	knownServiceCache.Set(hostname, known)
+	knownServiceCache.Set(hostname, known, gocache.DefaultExpiration)
 
 	return known
 }