… vendor/github.com/patrickmn/go-cache/LICENSE
… vendor/github.com/patrickmn/go-cache/cache.go
… vendor/github.com/patrickmn/go-cache/sharded.go
… vendor/manifest