`gcache` took an `interface{}`, so every time it is called a string header object gets allocated on the heap.
`go-cache` takes a `string` so avoids this allocation

Cache eviction is time-based so not as nice as before.

Benchmarks before (c5bdebdffebfa7cc0839770a3bf70429260f1bc9):
```
BenchmarkTopologyList-2         	       2	 603197223 ns/op	112161732 B/op	 1253503 allocs/op
BenchmarkTopologyHosts-2        	       3	 385415232 ns/op	82957173 B/op	 1015204 allocs/op
BenchmarkTopologyContainers-2   	      10	 192766565 ns/op	33145659 B/op	  407698 allocs/op
```

after this PR:
```
BenchmarkTopologyList-2         	       2	 528033687 ns/op	110843492 B/op	 1179523 allocs/op
BenchmarkTopologyHosts-2        	       3	 379481628 ns/op	81654378 B/op	  941519 allocs/op
BenchmarkTopologyContainers-2   	      10	 164464821 ns/op	32557828 B/op	  371446 allocs/op
```