@rade 2017-11-25
> Perhaps give https://github.com/gocql/gocql/blob/master/internal/lru/lru.go a try

That didn't work out because the package has "internal" in the path, which go refuses to import. I found a better implementation though, that includes locking. See #2952 for a PR replacing this one.